Binarias/Deriv/requirements.txt
# Dependências do Deriv Trading Bot
# Instalação: pip install -r requirements.txt

# WebSocket client (conexão com a API Deriv)
websocket-client>=1.6.0
